[vector][python] Read vindex vector indexes
