+++ paimon-python/dev/requirements-dev.txt
@@ -32,3 +32,5 @@ vortex-data==0.70.0; python_version >= "3.11"
 datafusion>=52; python_version >= "3.10"
 # Lumina vector search (optional, for lumina index tests)
 lumina-data>=0.1.0
+# paimon-vindex vector search (optional, for vindex index tests)
+paimon-vindex==0.1.0; python_version >= "3.9"
