+++ paimon-python/setup.py
@@ -180,6 +180,9 @@ setup(
         'lumina': [
             'lumina-data>=0.1.0'
         ],
+        'vindex': [
+            'paimon-vindex==0.1.0; python_version>="3.9"',
+        ],
         'sql': [
             'pypaimon-rust; python_version>="3.10"',
             'datafusion>=52; python_version>="3.10"',
